+++ agent/check_handler.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"github.com/sensu/sensu-go/agent/transformers"
+	"github.com/sensu/sensu-go/asset"
 	"github.com/sensu/sensu-go/command"
 	"github.com/sensu/sensu-go/transport"
 	"github.com/sensu/sensu-go/types"
@@ -68,9 +69,12 @@ func (a *Agent) executeCheck(request *types.CheckRequest) {
 		Check: check,
 	}
 
-	// Ensure that the asset manager is aware of all the assets required to
-	// execute the given check.
-	assets := a.assetManager.RegisterSet(checkAssets)
+	// Fetch and install all assets required for check execution.
+	assets, err := asset.GetAll(a.assetManager, checkAssets)
+	if err != nil {
+		a.sendFailure(event, fmt.Errorf("error getting assets for event: %s", err))
+		return
+	}
 
 	// Inject the dependenices into PATH, LD_LIBRARY_PATH & CPATH so that they are
 	// availabe when when the command is executed.
@@ -91,12 +95,6 @@ func (a *Agent) executeCheck(request *types.CheckRequest) {
 			return
 		}
 		ex.Input = string(input)
-	}
-
-	// Ensure that all the dependencies are installed.
-	if err := assets.InstallAll(); err != nil {
-		a.sendFailure(event, fmt.Errorf("error installing dependencies: %s", err))
-		return
 	}
 
 	checkExec, err := a.executor.Execute(context.Background(), ex)