+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 sensu/slack-handler.
 
 ### Changed
+- Changed sensu-agent's internal asset manager to use BoltDB.
 - Changed sensuctl title colour to use terminal's configured default for bold
 text.
 - The backend no longer forcibly binds to localhost.